Add error management in server arguments

+++ client/include/AUser.hpp
@@ -18,16 +18,16 @@
 class AUser : public CustomButton {
     Q_OBJECT
     public:
-    explicit AUser(QWidget *parent = nullptr, const std::string &login = "login", const std::string &ip = "ip", short port = -1, int id = -1) : CustomButton(parent, QString(login.c_str())), _login(login), _ip(ip), _port(port), _id(id) {};
+    explicit AUser(QWidget *parent = nullptr, const std::string &login = "login", const std::string &ip = "ip", int port = -1, int id = -1) : CustomButton(parent, QString(login.c_str())), _login(login), _ip(ip), _port(port), _id(id) {};
     virtual std::string getLogin() const = 0;
     virtual std::string getIP() const = 0;
-    virtual short getPort() const = 0;
+    virtual int getPort() const = 0;
     virtual int getID() const = 0;
 
 protected:
     std::string _login;
     std::string _ip;
-    short _port;
+    int _port;
     int _id;
 
 };
