+++ client/include/CustomMainWindow.hpp
@@ -48,7 +48,7 @@ private:
     int _otherId;
     std::string _otherIP;
     std::string _otherLogin;
-    short _otherPort = -1;
+    int _otherPort = -1;
 
     std::string _serverIP;
     std::string _serverPort;
